util: move GeoJSON coords / type logic in own module
- to be reuse in 'scattergeo'
